feat: moving wasm init into repective packages, updating node
