+++ packages/sdk-multicore/src/initThreadPool.ts
@@ -1,0 +1,1 @@
+export { initThreadPool } from "./../../wasm/src/sdk/sdk";
