+++ packages/sdk-node/src/index.ts
@@ -1,7 +1,7 @@
 import { webcrypto } from "node:crypto";
 import { Query as QueryWasm, Sdk as SdkWasm } from "@namada/wasm";
 // We have to use relative imports here othewise ts-patch is getting confused and produces wrong paths after compialtion
-import { init } from "../../wasm/src/init";
+import init from "./init";
 import { Sdk, SdkWasmOptions } from "../../lib/src";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -29,7 +29,7 @@ export async function initSdk(props: SdkWasmOptions): Promise<Sdk> {
   const wasm = await fetch("sdk.namada.wasm").then((wasm) =>
     wasm.arrayBuffer(),
   );
-  const { memory } = await init(wasm);
+  const { memory } = init();
 
   // Instantiate QueryWasm
   const query = new QueryWasm(rpcUrl, maspIndexerUrl);
