+++ packages/sdk/src/initInline.ts
@@ -1,6 +1,10 @@
+import initWasm, { InitOutput } from "../../wasm/src/sdk/sdk";
+// @ts-ignore
+import wasm from "../../wasm/src/sdk/sdk_bg.wasm?url";
+
+export const init: () => Promise<InitOutput> = async () => await initWasm(wasm);
 import { Query as QueryWasm, Sdk as SdkWasm } from "@namada/wasm";
 // We have to use relative imports here othewise ts-patch is getting confused and produces wrong paths after compilation
-import { init } from "../../wasm/src/init-inline";
 import { Sdk, SdkWasmOptions } from "../../lib/src";
 
 /**
