--- packages/wasm/src/init-thread-pool.ts
@@ -1,1 +1,0 @@
-export { initThreadPool } from "./sdk/sdk";
