--- packages/wasm/src/init.ts
@@ -1,5 +1,0 @@
-import initWasm, { InitInput, InitOutput } from "./sdk/sdk";
-
-export const init: (wasm: InitInput) => Promise<InitOutput> = async (wasm) => {
-  return await initWasm(wasm);
-};
